+++ src/CompoundCurve.php
@@ -165,6 +165,32 @@ class CompoundCurve extends Curve
         return new CompoundCurve($cs, ...$curves);
     }
 
+    public function withoutZ(): CompoundCurve
+    {
+        if (! $this->coordinateSystem->hasZ()) {
+            return $this;
+        }
+
+        $cs = $this->coordinateSystem->withZ(false);
+
+        $curves = array_map(fn(Curve $curve) => $curve->withoutZ(), $this->curves);
+
+        return new CompoundCurve($cs, ...$curves);
+    }
+
+    public function withoutM(): CompoundCurve
+    {
+        if (! $this->coordinateSystem->hasM()) {
+            return $this;
+        }
+
+        $cs = $this->coordinateSystem->withM(false);
+
+        $curves = array_map(fn(Curve $curve) => $curve->withoutM(), $this->curves);
+
+        return new CompoundCurve($cs, ...$curves);
+    }
+
     public function getBoundingBox() : BoundingBox
     {
         $boundingBox = new BoundingBox();