+++ src/Proxy/MultiSurfaceProxy.php
@@ -168,6 +168,24 @@ class MultiSurfaceProxy extends \Brick\Geo\MultiSurface implements ProxyInterfac
         return $this->proxyGeometry->toXY();
     }
 
+    public function withoutZ() : \Brick\Geo\GeometryCollection
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->withoutZ();
+    }
+
+    public function withoutM() : \Brick\Geo\GeometryCollection
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->withoutM();
+    }
+
     public function getBoundingBox() : \Brick\Geo\BoundingBox
     {
         if ($this->proxyGeometry === null) {