+++ src/Proxy/TINProxy.php
@@ -159,6 +159,24 @@ class TINProxy extends \Brick\Geo\TIN implements ProxyInterface
         return $this->proxyGeometry->toXY();
     }
 
+    public function withoutZ() : \Brick\Geo\PolyhedralSurface
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->withoutZ();
+    }
+
+    public function withoutM() : \Brick\Geo\PolyhedralSurface
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->withoutM();
+    }
+
     public function getBoundingBox() : \Brick\Geo\BoundingBox
     {
         if ($this->proxyGeometry === null) {